VCUP/requirements.txt
python-dotenv~=1.0.1
openpyxl~=3.1.5
requests~=2.32.3
json
re
urllib3